requirements.txt
Django==4.1.3
djangorestframework==3.15.1
djangorestframework==3.16.0
enums==0.0.2
Pillow==11.3.0
pytest==8.4.1
